Implement Repository Pattern & Integrate Hilt
* implemented the repository pattern to better handle data access before implementing caching & paging
* integrated hilt for a smother dependency injection
