+++ app/src/main/java/com/mis/route/newsapp/domain/repositories/ArticlesSourceRepository.kt
@@ -1,0 +1,7 @@
+package com.mis.route.newsapp.domain.repositories
+
+import com.mis.route.newsapp.data.data_sources.remote.news_api.models.sources.Source
+
+interface ArticlesSourceRepository {
+    suspend fun getSources(category: String): List<Source?>?
+}
