+++ app/src/main/java/com/mis/route/newsapp/presentations/application/NewsApplication.kt
@@ -1,0 +1,7 @@
+package com.mis.route.newsapp.presentations.application
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class NewsApplication : Application()
