+++ app/src/main/java/com/mis/route/newsapp/presentations/ui/home/HomeViewModel.kt
@@ -1,0 +1,5 @@
+package com.mis.route.newsapp.presentations.ui.home
+
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel : ViewModel()
