+++ app/src/main/java/com/mis/route/newsapp/presentations/ui/home/fragments/catergories/CategoriesFragment.kt
@@ -1,0 +1,45 @@
+package com.mis.route.newsapp.presentations.ui.home.fragments.catergories
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.mis.route.newsapp.databinding.FragmentCategoriesBinding
+
+class CategoriesFragment : Fragment() {
+    private lateinit var binding: FragmentCategoriesBinding
+    private lateinit var viewModel: CategoriesViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        viewModel = ViewModelProvider(this)[CategoriesViewModel::class.java]
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentCategoriesBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setObservers()
+    }
+
+    private fun setObservers() {
+        viewModel.category.observe(viewLifecycleOwner) {
+            onCategoryPicked.onPicked(it)
+        }
+    }
+
+    lateinit var onCategoryPicked: OnCategoryPicked
+
+    fun interface OnCategoryPicked {
+        fun onPicked(category: String)
+    }
+}
