+++ app/src/main/java/com/mis/route/newsapp/presentations/ui/home/fragments/news/MyGlideApp.kt
@@ -1,0 +1,7 @@
+package com.mis.route.newsapp.presentations.ui.home.fragments.news
+
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
+
+@GlideModule
+class MyGlideApp : AppGlideModule()
