--- app/src/main/java/com/mis/route/newsapp/ui/home/HomeViewModel.kt
@@ -1,5 +1,0 @@
-package com.mis.route.newsapp.ui.home
-
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel()
