--- app/src/main/java/com/mis/route/newsapp/ui/home/fragments/news/DialogMessage.kt
@@ -1,13 +1,0 @@
-package com.mis.route.newsapp.ui.home.fragments.news
-
-import android.content.DialogInterface
-
-data class DialogMessage(
-    val title: String?,
-    val message: String?,
-    val posText: String?,
-    val posAction: DialogInterface.OnClickListener?,
-    val negText: String?,
-    val negAction: DialogInterface.OnClickListener?,
-    val cancelable: Boolean = true
-)
