--- app/src/main/java/com/mis/route/newsapp/ui/home/fragments/settings/SettingsFragment.kt
@@ -1,21 +1,0 @@
-package com.mis.route.newsapp.ui.home.fragments.settings
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.mis.route.newsapp.databinding.FragmentSettingsBinding
-
-class SettingsFragment : Fragment() {
-    private lateinit var binding: FragmentSettingsBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentSettingsBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-}
