--- app/src/main/java/com/mis/route/newsapp/webservices/ApiManager.kt
@@ -1,20 +1,0 @@
-package com.mis.route.newsapp.webservices
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object ApiManager {
-    private var retrofit: Retrofit? = null
-
-    private fun getInstance(): Retrofit {
-        if (retrofit == null) {
-            retrofit = Retrofit.Builder()
-                .baseUrl("https://newsapi.org/")
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-        }
-        return retrofit!!
-    }
-
-    fun getApi(): WebServices = getInstance().create(WebServices::class.java)
-}
