+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/main_background"
-    tools:context=".ui.splash.SplashActivity">
+    tools:context=".presentations.ui.splash.SplashActivity">
 
     <ImageView
         android:layout_width="wrap_content"
