+++ build.gradle
@@ -3,4 +3,5 @@ plugins {
     id 'com.android.application' version '8.0.1' apply false
     id 'com.android.library' version '8.0.1' apply false
     id 'org.jetbrains.kotlin.android' version '1.8.20' apply false
+    id("com.google.dagger.hilt.android") version "2.51.1" apply false
 }
